news_api/requirements.txt
Django==3.2.19
gunicorn==20.1.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
